Fix player performance/math bugs
Prior to the commit the demo would work for a bit then audio underflows would
start and increase over time. This fixes a number of bugs in
AudioRenderer.fillDataBuffer(). The big ones being
* inBuffer computation mixed units of time and elements, causing us to
always fill far below target
* when inBuffer appeared >= target, we neglected to return; so we still
scheduled a bunch of reads that somewhat hid the issue
* we scheduled more setTimeout(...) work than needed, including when not
playing.

I also noticed that we weren't checking that ring buffer actually wrote
all that we requested. Turns out we were ok, but we were getting a
little close to full (85%), so I cranked up the DATA_BUFFER_DURATION a
bit. We should probably have an overflow buffer to avoid data loss.

There's still a lot of low hanging fruit for performance improvements:
1. use a common demuxer between audio/video
2. move video and audio rendering to workers

Will send a follow up for those^ next week.

Aside: I still hear the occasional audio glitch. It's not the file, and
it's not an underflow. Will be interesting to see if the additional
improvements remedy that.